[FE][Refactor][Desgin] commu redux로 수정 및 CSS폴더로 스타일 컴포넌트 이동
