+++ client/src/page/community/Community.js
@@ -1,115 +1,11 @@
 import Layout from '../../common/Layout';
-import styled from 'styled-components';
-import Commpost from './Commpost';
-import React, { useEffect } from 'react';
+import { CommunityDesign } from '../../style/CommunityDesign';
+import CommunityPost from './CommunityPost';
+import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import { commulist } from '../../redux/boardSlice';
-import { useSelector, useDispatch } from 'react-redux';
-
-const CommunityDesign = styled.div`
-  margin: 0;
-  padding: 0;
-  border: 0;
-  font-size: 100%;
-  font: inherit;
-  vertical-align: baseline;
-  background-color: #f9e6e6;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  font-family: 'Noto Sans KR', sans-serif;
-
-  .flexcontent {
-    display: flex;
-    justify-content: space-between;
-    flex-direction: wrap;
-
-    .community {
-      text-align: left;
-      font-size: 32px;
-      margin-bottom: 15px;
-      margin-top: 15px;
-      margin-left: 15px;
-      font-family: 'MaplestoryOTFBold';
-    }
-    .writebutton {
-      background-color: #f05858;
-      border-radius: 3px;
-      border: none;
-      color: white;
-      width: 60px;
-      height: 40px;
-      margin-bottom: 15px;
-      margin-top: 15px;
-      cursor: pointer;
-      text-align: right;
-      margin-right: 25px;
-      font-family: 'MaplestoryOTFBold';
-      &:hover {
-        background-color: #d32f2f;
-      }
-    }
-  }
-
-  ul {
-    list-style: none;
-    display: flex;
-    justify-content: space-around;
-    text-align: center;
-    align-items: center;
-    margin-top: 2px;
-    padding: 15px 10px;
-    border-bottom: 1px solid #d6d6d6;
-    font-weight: 500;
-
-    .infoid {
-      flex: 0.2;
-    }
-    .infoperson {
-      flex: 0.4;
-    }
-    .infotitle {
-      flex: 1;
-    }
-    .infoview {
-      flex: 0.4;
-    }
-    .infocreat {
-      flex: 0.3;
-    }
-  }
-
-  .allcomm {
-    justify-content: center;
-    background-color: white;
-    width: 70vw;
-    height: 80vh;
-  }
-
-  .writepost {
-    font-size: 18px;
-  }
-  .writeinfo {
-    overflow: hidden;
-    white-space: nowrap;
-    display: flex;
-    justify-content: space-around;
-  }
-`;
 
 const Community = () => {
   const Navigate = useNavigate();
-  const dispatch = useDispatch();
-  const data = useSelector((state) => state.commulist.data);
-
-  useEffect(() => {
-    {
-      dispatch(commulist());
-    }
-  }, [dispatch]);
 
   const goWrite = () => {
     Navigate('/commu/posts');
@@ -134,7 +30,7 @@ const Community = () => {
             <li className="infoview">조회수</li>
             <li className="infocreat">작성시간</li>
           </ul>
-          <Commpost data={data} />
+          <CommunityPost />
         </div>
       </CommunityDesign>
     </Layout>
