+++ client/src/page/survey/SurveyStep/Step5.js
@@ -1,70 +1,7 @@
 import React from 'react';
-import styled from 'styled-components';
 import high from '../../../../public/high.jpg';
 import logo from '../../../../public/logo.png';
-
-const SurveyResultDesign = styled.div`
-  font-family: 'Noto Sans KR', sans-serif;
-  display: flex;
-  flex-direction: column;
-  width: 100vw;
-  height: 100vh;
-  background-color: #f1f2f3;
-  font-size: 22px;
-
-  img {
-    width: 150px;
-    height: 50px;
-  }
-  .whoresult {
-    border-bottom: 1px solid black;
-    margin-top: 10vh;
-  }
-
-  div span {
-    margin-left: 5vw;
-  }
-  .explain {
-    margin-top: 7vh;
-    margin-left: 5vw;
-  }
-  .explaindetail {
-    width: 80vw;
-    margin-top: 5vh;
-  }
-  .helppill {
-    margin-top: 5vh;
-    margin-left: 1vw;
-  }
-
-  .recommend {
-    background-color: #f5d6d6;
-    margin-top: 5vh;
-    margin-left: 5vw;
-    width: 80vw;
-    height: 50vh;
-    flex-wrap: wrap;
-  }
-  .efficacy {
-    margin-top: 5vh;
-  }
-
-  .buynow {
-    margin-left: 2vw;
-    margin-top: 20px;
-  }
-  .retry {
-    background-color: #f05858;
-    border: none;
-    margin-left: 5vw;
-    width: 160px;
-    height: 70px;
-    font-size: 20px;
-    color: white;
-  }
-  .high {
-  }
-`;
+import { SurveyResultDesign } from '../../../style/SurveyStyle';
 
 const Step5 = (/*{form}*/) => {
   // 소화 , 피부, 눈 , 면역, 피로
