+++ client/src/redux/boardSlice.js
@@ -1,21 +1,10 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 import axios from 'axios';
+import { GetCommulist } from './CommuntiySlice';
+import { useSelector, useDispatch } from 'react-redux';
 
 const API_SERVER = process.env.API_SERVER;
 // API_SERVER
-
-//commu 전체 조회하기
-export const commulist = createAsyncThunk(async () => {
-  try {
-    const response = await axios.get(`http://localhost:3000/commu`, {
-      withCredentials: true,
-    });
-    return response.data;
-  } catch (error) {
-    console.error(`http://localhost:3000/commu`, error);
-    throw error;
-  }
-});
 
 //userid는 로그인할 때 받아오기
 export const submitPost = createAsyncThunk(
@@ -33,6 +22,8 @@ export const submitPost = createAsyncThunk(
           withCredentials: true,
         },
       );
+      const dispatch = useDispatch();
+      dispatch(GetCommulist());
     } catch (error) {
       console.error(`${API_SERVER}/commu/posts`, error);
     }
@@ -60,7 +51,7 @@ export const updatePost = createAsyncThunk(
   'board/updatePost',
   async ({ commuId, title, content }) => {
     try {
-      await axios.patch(
+      const response = await axios.patch(
         `${API_SERVER}/commu/${commuId}`,
         {
           title,
@@ -70,6 +61,7 @@ export const updatePost = createAsyncThunk(
           withCredentials: true,
         },
       );
+      return response.data;
     } catch (error) {
       console.error(`${API_SERVER}/commu/${commuId}`, error);
     }
