+++ client/src/redux/store.js
@@ -1,8 +1,9 @@
 import { configureStore } from '@reduxjs/toolkit';
 import boardSlice from './boardSlice';
+import CommunitySlice from './CommuntiySlice';
 
 const store = configureStore({
-  reducer: { board: boardSlice },
+  reducer: { board: boardSlice, commu: CommunitySlice },
 });
 
 export default store;
